Updated User view & controller
- Amended user view
- Amended user controller
- quick amend to user migration for is client user an admin

+++ resources/views/admin/users/create_edit.blade.php
@@ -24,78 +24,204 @@
 			<input type="file" multiple="true" class="form-control" name="avatar" id="avatar" />
 		</div>
 
-		<div class="panel panel-primary" data-collapsed="0">
-			<div class="panel-heading">
-				<div class="panel-title">
-					Username
+		<div class="row">
+
+			<div class="col-sm-4">
+
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">
+							Username
+						</div>
+						<div class="panel-options">
+							<a href="#" data-rel="collapse">
+								<i class="fa fa-angle-down"></i>
+							</a>
+						</div>
+					</div>
+
+					<div class="panel-body" style="display: block;">
+						@if($errors->first('username'))
+							<div class="alert alert-danger">
+								<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+								{{ $errors->first('username') }}
+							</div>
+						@endif
+						<p>User's Username</p>
+						<input type="text" class="form-control" name="username" id="username" autocomplete="off" value="{{
+						($user) ? $user->username : old('username')
+						}}" />
+					</div>
 				</div>
-				<div class="panel-options">
-					<a href="#" data-rel="collapse">
-						<i class="fa fa-angle-down"></i>
-					</a>
-				</div>
+
 			</div>
 
-			<div class="panel-body" style="display: block;">
-				@if($errors->first('username'))
-					<div class="alert alert-danger">
-						<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
-						{{ $errors->first('username') }}
+			<div class="col-sm-4">
+
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">
+							Full Name
+						</div>
+						<div class="panel-options">
+							<a href="#" data-rel="collapse">
+								<i class="fa fa-angle-down"></i>
+							</a>
+						</div>
 					</div>
-				@endif
-				<p>User's Username</p>
-				<input type="text" class="form-control" name="username" id="username" autocomplete="off" value="{{
-				($user) ? $user->username : old('username')
-				}}" />
+
+					<div class="panel-body" style="display: block;">
+						@if($errors->first('full_name'))
+							<div class="alert alert-danger">
+								<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+								{{ $errors->first('full_name') }}
+							</div>
+						@endif
+						<p>User's Full Name</p>
+						<input type="text" class="form-control" name="full_name" id="full_name" autocomplete="off" value="{{
+						($user) ? $user->full_name : old('full_name')
+						}}" />
+					</div>
+				</div>
+
 			</div>
+
+			<div class="col-sm-4">
+
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">
+							Job Title
+						</div>
+						<div class="panel-options">
+							<a href="#" data-rel="collapse">
+								<i class="fa fa-angle-down"></i>
+							</a>
+						</div>
+					</div>
+
+					<div class="panel-body" style="display: block;">
+						@if($errors->first('job_title'))
+							<div class="alert alert-danger">
+								<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+								{{ $errors->first('job_title') }}
+							</div>
+						@endif
+						<p>User's Job Title</p>
+						<input type="text" class="form-control" name="job_title" id="job_title" autocomplete="off" value="{{
+						($user) ? $user->job_title : old('job_title')
+						}}" />
+					</div>
+				</div>
+
+			</div>
+
 		</div>
 
-		<div class="panel panel-primary" data-collapsed="0">
-			<div class="panel-heading">
-				<div class="panel-title">Email</div>
+		<div class="row">
 
-				<div class="panel-options">
-					<a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a>
-				</div>
-			</div>
+			<div class="col-sm-4">
 
-			<div class="panel-body" style="display: block;">
-				@if($errors->first('email'))
-					<div class="alert alert-danger">
-						<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
-						{{ $errors->first('email') }}
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">Email</div>
+
+						<div class="panel-options">
+							<a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a>
+						</div>
 					</div>
-				@endif
 
-				<p>User's Email Address</p>
-				<input type="text" class="form-control" name="email" id="email" autocomplete="off" value="{{
-				($user) ? $user->email : old('email')
-				}}" readonly onfocus="this.removeAttribute('readonly');"/>
+					<div class="panel-body" style="display: block;">
+						@if($errors->first('email'))
+							<div class="alert alert-danger">
+								<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+								{{ $errors->first('email') }}
+							</div>
+						@endif
+
+						<p>User's Email Address</p>
+						<input type="text" class="form-control" name="email" id="email" autocomplete="off" value="{{
+						($user) ? $user->email : old('email')
+						}}" readonly onfocus="this.removeAttribute('readonly');"/>
+					</div>
+				</div>
+
 			</div>
+
+			<div class="col-sm-4">
+
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">
+							Phone Number
+						</div>
+						<div class="panel-options">
+							<a href="#" data-rel="collapse">
+								<i class="fa fa-angle-down"></i>
+							</a>
+						</div>
+					</div>
+
+					<div class="panel-body" style="display: block;">
+						@if($errors->first('tel'))
+							<div class="alert alert-danger">
+								<button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+								{{ $errors->first('tel') }}
+							</div>
+						@endif
+						<p>User's Phone Number</p>
+						<input type="text" class="form-control" name="tel" id="tel" autocomplete="off" value="{{
+						($user) ? $user->tel : old('tel')
+						}}" />
+					</div>
+				</div>
+
+			</div>
+
 		</div>
 
-		<div class="panel panel-primary" data-collapsed="0">
-			<div class="panel-heading">
-				<div class="panel-title">Password</div>
-				<div class="panel-options">
-					<a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a>
+		<div class="row">
+
+			<div class="col-sm-4">
+
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">Password</div>
+						<div class="panel-options">
+							<a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a>
+						</div>
+					</div>
+
+					<div class="panel-body" style="display: block;">
+		                @if($errors->first('password'))
+		                    <div class="alert alert-danger">
+		                        <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+		                        {{ $errors->first('password') }}
+		                    </div>
+		                @endif
+
+		                <p>
+		                    {{ ($user) ? '(leave empty to keep your original password)' : 'Enter users password:' }}
+		                </p>
+						<input type="password" class="form-control" name="password" id="password" autocomplete="off" value=""
+		                       readonly onfocus="this.removeAttribute('readonly');"  title="password"/>
+					</div>
 				</div>
+
 			</div>
 
-			<div class="panel-body" style="display: block;">
-                @if($errors->first('password'))
-                    <div class="alert alert-danger">
-                        <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
-                        {{ $errors->first('password') }}
-                    </div>
-                @endif
+			<div class="col-sm-4">
 
-                <p>
-                    {{ ($user) ? '(leave empty to keep your original password)' : 'Enter users password:' }}
-                </p>
-				<input type="password" class="form-control" name="password" id="password" autocomplete="off" value="" 
-                       readonly onfocus="this.removeAttribute('readonly');"  title="password"/>
+				<div class="panel panel-primary" data-collapsed="0"> <div class="panel-heading">
+					<div class="panel-title">Active</div> <div class="panel-options"> <a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a> </div></div>
+					<div class="panel-body" style="display: block;">
+						<label>Active</label>
+						<input type="checkbox" id="active" name="active" {{ ((($user) && ($user->active)) || (!$user)) ? 'checked="checked" value=1' : '' }} />
+					</div>
+				</div>
+
 			</div>
+
 		</div>
 
 		<div class="row">
@@ -130,6 +256,7 @@
 			</div>
 
 			<div class="col-sm-4" id="client-box">
+
 				<div class="panel panel-primary" data-collapsed="0">
 					<div class="panel-heading">
 						<div class="panel-title">Client</div>
@@ -159,17 +286,30 @@
 						</select>
 					</div>
 				</div>
+
 			</div>
 
-			<div class="col-sm-4">
-				<div class="panel panel-primary" data-collapsed="0"> <div class="panel-heading">
-					<div class="panel-title">Active</div> <div class="panel-options"> <a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a> </div></div>
+			<div class="col-sm-4" id="client-admin">
+
+				<div class="panel panel-primary" data-collapsed="0">
+					<div class="panel-heading">
+						<div class="panel-title">Is this user an Admin</div>
+
+						<div class="panel-options">
+							<a href="#" data-rel="collapse"><i class="fa fa-angle-down"></i></a>
+						</div>
+					</div>
+
 					<div class="panel-body" style="display: block;">
-						<label>Active</label>
-						<input type="checkbox" id="active" name="active" {{ ((($user) && ($user->active)) || (!$user)) ? 'checked="checked" value=1' : '' }} />
+						<select id="is_client_admin" name="is_client_admin">
+							<option value="0" @if(isset($user->is_client_admin)&&$user->is_client_admin==0) selected @endif>No</option>
+							<option value="1" @if(isset($user->is_client_admin)&&$user->is_client_admin==1) selected @endif>Yes</option>				
+						</select>
 					</div>
 				</div>
+
 			</div>
+
 		</div>
 
 		@if(isset($user->id))
@@ -181,12 +321,14 @@
 		<input type="submit" value="{{ ($user) ? 'Update' : 'Create' }} User" class="btn btn-success pull-right" />
 	</form>
 
-	{!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id], 'id' => 'form-delete-users-' . $user->id]) !!}
-    <a href="" class="btn btn-danger delete" data-form="users-{{ $user->id }}">
-        <i class="fa fa-trash-o"></i>
-        Delete
-    </a>
-    {!! Form::close() !!}
+	@if(isset($user->id))
+		{!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id], 'id' => 'form-delete-users-' . $user->id]) !!}
+	    <a href="" class="btn btn-danger delete" data-form="users-{{ $user->id }}">
+	        <i class="fa fa-trash-o"></i>
+	        Delete
+	    </a>
+	    {!! Form::close() !!}
+	@endif
 </div>
 @endsection
 @section('javascript')
@@ -202,8 +344,17 @@
 				}
 			});
 
+			$('#role').change(function(){
+				if($(this).val() == 'client'){
+					$('#client-admin').show();
+				} else {
+					$('#client-admin').hide();
+				}
+			});
+
 			if($('#role').val() != 'client'){
 				$('#client-box').hide();
+				$('#client-admin').hide();
 			}
 
 			$('#active, #disabled').change(function() {
